Fix context menu in portfolio find
