+++ BlockSettleUILib/PortfolioWidget.h
@@ -15,7 +15,7 @@
 #include <QMenu>
 #include <memory>
 
-#include "TabWithShortcut.h"
+#include "TransactionsWidgetInterface.h"
 
 namespace spdlog {
    class logger;
@@ -37,11 +37,11 @@ class ArmoryConnection;
 class CCPortfolioModel;
 class MarketDataProvider;
 class MDCallbacksQt;
-class SignContainer;
+class WalletSignerContainer;
 class TransactionsViewModel;
 class UnconfirmedTransactionFilter;
 
-class PortfolioWidget : public TabWithShortcut
+class PortfolioWidget : public TransactionsWidgetInterface
 {
 Q_OBJECT
 
@@ -55,7 +55,7 @@ public:
       , const std::shared_ptr<MarketDataProvider> &
       , const std::shared_ptr<MDCallbacksQt> &
       , const std::shared_ptr<CCPortfolioModel> &
-      , const std::shared_ptr<SignContainer> &
+      , const std::shared_ptr<WalletSignerContainer> &
       , const std::shared_ptr<ArmoryConnection> &
       , const std::shared_ptr<bs::UTXOReservationManager> &utxoReservationManager
       , const std::shared_ptr<spdlog::logger> &
@@ -69,26 +69,9 @@ private slots:
    void showTransactionDetails(const QModelIndex& index);
    void showContextMenu(const QPoint& point);
 
-   void onCreateRBFDialog();
-   void onCreateCPFPDialog();
-
 private:
    std::unique_ptr<Ui::PortfolioWidget> ui_;
-   std::shared_ptr<TransactionsViewModel> model_;
-   std::shared_ptr<spdlog::logger>        logger_;
-   std::shared_ptr<ApplicationSettings>   appSettings_;
    UnconfirmedTransactionFilter* filter_;
-   QMenu    contextMenu_;
-   QAction  *actionRBF_;
-   QAction  *actionCPFP_;
-   QAction  *actionCopyAddr_;
-   QAction  *actionCopyTx_;
-   QString  curAddress_;
-   QString  curTx_;
-   std::shared_ptr<SignContainer>      signContainer_;
-   std::shared_ptr<ArmoryConnection>   armory_;
-   std::shared_ptr<bs::sync::WalletsManager> walletsManager_;
-   std::shared_ptr<bs::UTXOReservationManager> utxoReservationManager_;
 };
 
 #endif // __PORFOLIO_WIDGET_H__
